[cleanup] Clean up old connectors.
